pyproject.toml
[tool.poetry]
name = "servicenow-doc-template"
version = "0.1.0"
description = "ServiceNow ITSM導入プロジェクト向け設計書テンプレート生成ツール"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
jinja2 = "^3.1.2"
click = "^8.1.7"
pydantic = "^2.5.0"
python-docx = "^1.1.0"
pyyaml = "^6.0.1"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
isort = "^5.12.0"

[tool.poetry.scripts]
servicenow-doc = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"